+++ sql-plugin/src/main/spark311/scala/org/apache/spark/sql/rapids/aggregate/AggregateFunctions.scala
@@ -31,11 +31,12 @@
 {"spark": "332"}
 {"spark": "333"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql.rapids
+package org.apache.spark.sql.rapids.aggregate
 
 import com.nvidia.spark.rapids.{GpuCast, GpuWindowExpression, GpuWindowSpecDefinition}
 
 import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.rapids.GpuDecimalDivide
 import org.apache.spark.sql.types.DecimalType
 
 abstract class GpuDecimalAverage(child: Expression, sumDataType: DecimalType)
