+++ sql-plugin/src/main/spark330db/scala/org/apache/spark/sql/rapids/aggregate/AggregateFunctions.scala
@@ -21,11 +21,12 @@
 {"spark": "341"}
 {"spark": "350"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql.rapids
+package org.apache.spark.sql.rapids.aggregate
 
 import com.nvidia.spark.rapids.{GpuCast, GpuWindowExpression, GpuWindowSpecDefinition}
 
 import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.rapids.GpuDecimalDivide
 import org.apache.spark.sql.types.DecimalType
 
 abstract class GpuDecimalAverage(child: Expression, sumDataType: DecimalType)
