+++ tests/src/test/scala/com/nvidia/spark/rapids/HashAggregateRetrySuite.scala
@@ -24,7 +24,7 @@ import com.nvidia.spark.rapids.jni.RmmSpark
 import org.mockito.Mockito._
 import org.scalatestplus.mockito.MockitoSugar
 
-import org.apache.spark.sql.rapids.{CudfAggregate, CudfSum}
+import org.apache.spark.sql.rapids.aggregate.{CudfAggregate, CudfSum}
 import org.apache.spark.sql.types.{DataType, IntegerType, LongType}
 
 class HashAggregateRetrySuite
